pyproject.toml
[tool.poetry]
name = "document-to-markdown"
version = "0.1.0"
description = ""
authors = ["drmingler <davidemmanuel75@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
docling = "^2.3.1"
python-multipart = "^0.0.17"
celery = "^5.4.0"
flower = "^2.0.1"
redis = "^5.2.0"
gunicorn = "^23.0.0"
supabase = "^2.10.0"
sentence-transformers = "^3.3.0"
pyjwt = "^2.10.0"
httpx = "^0.27.2"
tenacity = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
